Massive update
AI updates, UI on startup, fixed lag.

+++ Evo.Core/Interfaces/IMovable.cs
@@ -10,9 +10,9 @@ namespace Evo.Core.Interfaces
     public interface IMovable
     {
         object Target { get; set; }
-        int Speed { get; set; }
-        int MaxSpeed { get; set; }
-        int MinSpeed { get; set; }
+        double Speed { get; set; }
+        double MaxSpeed { get; set; }
+        double MinSpeed { get; set; }
         Vector2 Direction { get; set; }
         void Move();
         void Rotate(float angle);
